% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance.R
\name{deviance.SLOPE}
\alias{deviance.SLOPE}
\title{Model deviance}
\usage{
\method{deviance}{SLOPE}(object, ...)
}
\arguments{
\item{object}{an object of class \code{'SLOPE'}.}

\item{...}{ignored}
}
\value{
For Gaussian models this is twice the residual sums of squares. For
all other models, two times the negative loglikelihood is returned.
}
\description{
Model deviance
}
\examples{
fit <- SLOPE(abalone$x, abalone$y, family = "poisson")
deviance(fit)
}
