% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_path.R
\name{flow_path}
\alias{flow_path}
\title{Flow Path}
\usage{
flow_path(x, y)
}
\arguments{
\item{x}{Input DEM raster}

\item{y}{Input streams raster}
}
\description{
Performs the flow path in DEMs based on the wbt_downslope_distance_to_stream() function (Lindsay, 2016) for calculates sediment connectivity (Cavalli et al. 2013).
}
\examples{
\dontrun{
library(terra)
library(slope)
dem <- terra::rast(system.file("ex/elev.tif", package = "terra"))
streams <- terra::rast(system.file("ex/streams.tif", package = "terra"))
fp <- slope::flow_path(dem, s)
plot(fp)
}
}
