% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillshade.R
\name{hillshade}
\alias{hillshade}
\title{Hillshade}
\usage{
hillshade(x, w, angle, direction)
}
\arguments{
\item{x}{Input DEM raster}

\item{w}{Number of cells for the window size}

\item{angle}{Illumination angle (altitude) in degrees}

\item{direction}{Illumination direction (azimuth) in degrees}
}
\description{
Calculates shaded relief (hillshade) from a Digital Elevation Model (DEM).
}
\examples{
\dontrun{
library(terra)
library(slope)
dem <- terra::rast(system.file("ex/elev.tif", package = "terra"))
shade <- slope::hillshade(dem, 3, 45, 315)
plot(shade)
}
}
