% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolatePenalty.R
\name{interpolatePenalty}
\alias{interpolatePenalty}
\title{Interpolate penalty values}
\usage{
interpolatePenalty(penalty, x)
}
\arguments{
\item{penalty}{current penalty sequence}

\item{x}{new sequence}
}
\value{
Interpolated values of lambda
}
\description{
Interpolate penalty values
}
\author{
Jerome Friedman, Trevor Hastie, Rob Tibshirani, and Noah Simon
}
\keyword{internal}
