% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{create_lambda}
\alias{create_lambda}
\title{Lambda sequences for SLOPE (deprecated)}
\usage{
create_lambda(n, p, fdr = 0.2, method = c("bhq", "gaussian"))
}
\arguments{
\item{n}{number of observations}

\item{p}{number of variables}

\item{fdr}{target False Discovery Rate (FDR)}

\item{method}{method to use for computing \eqn{\lambda} (see Details)}
}
\description{
Computes \eqn{\lambda} sequences for SLOPE according to several pre-defined methods.
}
\details{
The following methods for computing \eqn{\lambda} are supported:
\itemize{
\item \code{bhq}: Computes sequence inspired by Benjamini-Hochberg (BHq)
procedure
\item \code{gaussian}: Computes modified BHq sequence inspired by
Gaussian designs
}
}
