% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_factor.R
\name{ls_factor}
\alias{ls_factor}
\title{LS Factor}
\usage{
ls_factor(x, type, method, crs, m, n)
}
\arguments{
\item{x}{Input DEM raster}

\item{type}{Flow accumulation calculation type: 'cells', 'sca', or 'ca'.}

\item{method}{[0] for Moore and Burch (1986) or [1] for Bertoni and Lombardi Neto (1985).}

\item{crs}{Coordinate reference systems}

\item{m}{Exponent of first equation}

\item{n}{Exponent of second equation}
}
\description{
Calculates LS Factor based in Bertoni and Lombardi Neto (1985) or Moore and Burch (1986).
}
\examples{
\dontrun{
library(terra)
library(slope)
dem <- terra:rast(system.file("ex/elev.tif", package = "terra"))
ls <- slope::ls_factor(dem, 'sca')
plot(ls)
}
}
