% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalization}
\alias{normalization}
\title{Transformation (Normalization)}
\usage{
normalization(x, w)
}
\arguments{
\item{x}{Input DEM raster}

\item{w}{Number of cells for the window size}
}
\description{
Normalization of Land Surface Parameters (LSPs), based Huang et al. (2021)
}
\examples{
\dontrun{
library(terra)
library(slope)
dem <- terra:rast(system.file("ex/elev.tif", package = "terra"))
ni <- slope::normalization(dem, 7)
plot(ni)
}
}
