% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Plot cluster structure}
\usage{
plot_clusters(
  x,
  plot_signs = FALSE,
  color_clusters = TRUE,
  include_zeroes = TRUE,
  show_alpha = TRUE
)
}
\arguments{
\item{x}{an object of class \code{'SLOPE'}}

\item{plot_signs}{logical, indicating whether to plot signs of estimated
coefficients on the plot}

\item{color_clusters}{logical, indicating whether the clusters should have
different colors}

\item{include_zeroes}{logical, indicating whether zero variables should be
plotted. Default to TRUE}

\item{show_alpha}{logical, indicatiung whether labels with alpha values or
steos in path should be plotted. Default TRUE}
}
\description{
Plot cluster structure
}
\examples{
set.seed(10)
X <- matrix(rnorm(10000), ncol = 10)
colnames(X) <- paste0("X", 1:10)
beta <- c(rep(10, 3), rep(-20, 2), rep(20, 2), rep(0, 3))
Y <- X \%*\% beta + rnorm(1000)
fit <- SLOPE(X, Y, patterns = TRUE)

plot_clusters(fit)
}
\seealso{
\code{\link[=SLOPE]{SLOPE()}}
}
