% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotClusters}
\alias{plotClusters}
\title{Plot cluster structure}
\usage{
plotClusters(
  x,
  plot_signs = FALSE,
  color_clusters = TRUE,
  include_zeroes = TRUE,
  show_alpha = FALSE,
  alpha_steps = NULL
)
}
\arguments{
\item{x}{an object of class \code{'SLOPE'}}

\item{plot_signs}{logical, indicating whether to plot signs of estimated
coefficients on the plot}

\item{color_clusters}{logical, indicating whether the clusters should have
different colors}

\item{include_zeroes}{logical, indicating whether zero variables should be
plotted. Default to TRUE}

\item{show_alpha}{logical, indicatiung whether labels with alpha values or
steps in the path should be plotted. Default FALSE.}

\item{alpha_steps}{a vector of integer alpha steps to plot. If NULL, all the steps
are plotted. Default to NULL.}
}
\value{
Invisibly returns NULL. The function is called for its
side effect of producing a plot.
}
\description{
Plot cluster structure
}
\examples{
set.seed(10)
X <- matrix(rnorm(10000), ncol = 10)
colnames(X) <- paste0("X", 1:10)
beta <- c(rep(10, 3), rep(-20, 2), rep(20, 2), rep(0, 3))
Y <- X \%*\% beta + rnorm(1000)
fit <- SLOPE(X, Y, patterns = TRUE)

plotClusters(fit)
plotClusters(fit, alpha_steps = 1:10)
}
\seealso{
\code{\link[=SLOPE]{SLOPE()}}
}
