% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ABSLOPE}
\alias{print.ABSLOPE}
\title{Print results from ABSLOPE fit}
\usage{
\method{print}{ABSLOPE}(x, ...)
}
\arguments{
\item{x}{an object of class \code{'ABSLOPE'}}

\item{...}{other arguments passed to \code{\link[=print]{print()}}}
}
\value{
Prints output on the screen
}
\description{
Print results from ABSLOPE fit
}
\examples{
set.seed(17)
xy <- SLOPE:::randomProblem(1e2, 2, response = "gaussian")
x <- as.matrix(xy$x)
y <- xy$y
fit <- ABSLOPE(x, y)
print(fit)

}
\seealso{
\code{\link[=ABSLOPE]{ABSLOPE()}}, \code{\link[=print.ABSLOPE]{print.ABSLOPE()}}

Other SLOPE-methods: 
\code{\link{coef.SLOPE}()},
\code{\link{deviance.SLOPE}()},
\code{\link{plot.SLOPE}()},
\code{\link{predict.SLOPE}()},
\code{\link{print.SLOPE}()},
\code{\link{score}()}
}
\concept{SLOPE-methods}
