% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.SLOPE}
\alias{print.SLOPE}
\alias{print.TrainedSLOPE}
\title{Print results from SLOPE fit}
\usage{
\method{print}{SLOPE}(x, ...)

\method{print}{TrainedSLOPE}(x, ...)
}
\arguments{
\item{x}{an object of class \code{'SLOPE'} or \code{'TrainedSLOPE'}}

\item{...}{other arguments passed to \code{\link[=print]{print()}}}
}
\value{
Prints output on the screen
}
\description{
Print results from SLOPE fit
}
\examples{
fit <- SLOPE(wine$x, wine$y, family = "multinomial")
print(fit, digits = 1)
}
\seealso{
\code{\link[=SLOPE]{SLOPE()}}, \code{\link[=print.SLOPE]{print.SLOPE()}}

Other SLOPE-methods: 
\code{\link{coef.SLOPE}()},
\code{\link{deviance.SLOPE}()},
\code{\link{plot.SLOPE}()},
\code{\link{predict.SLOPE}()},
\code{\link{score}()}
}
\concept{SLOPE-methods}
