% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary_SLOPE}
\alias{print.summary_SLOPE}
\title{Print Summary of SLOPE Model}
\usage{
\method{print}{summary_SLOPE}(x, digits = 3, ...)
}
\arguments{
\item{x}{an object of class \code{'summary_SLOPE'}}

\item{digits}{number of significant digits to print}

\item{...}{other arguments passed to \code{\link[=print]{print()}}}
}
\value{
Invisibly returns the input object
}
\description{
Print Summary of SLOPE Model
}
\seealso{
\code{\link[=summary.SLOPE]{summary.SLOPE()}}
}
