% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary_TrainedSLOPE}
\alias{print.summary_TrainedSLOPE}
\title{Print Summary of TrainedSLOPE Model}
\usage{
\method{print}{summary_TrainedSLOPE}(x, digits = 3, ...)
}
\arguments{
\item{x}{an object of class \code{'summary_TrainedSLOPE'}}

\item{digits}{number of significant digits to print}

\item{...}{other arguments passed to \code{\link[=print]{print()}}}
}
\value{
Invisibly returns the input object
}
\description{
Print Summary of TrainedSLOPE Model
}
\seealso{
\code{\link[=summary.TrainedSLOPE]{summary.TrainedSLOPE()}}
}
