% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{prox_sorted_L1}
\alias{prox_sorted_L1}
\title{Prox for sorted L1 norm (deprecated)}
\usage{
prox_sorted_L1(x, lambda, method)
}
\arguments{
\item{x}{input vector}

\item{lambda}{vector of \eqn{\lambda}'s, sorted in decreasing order}

\item{method}{deprecated}
}
\description{
Compute the prox for the sorted L1 norm. That is, given a vector \eqn{x}
and a decreasing vector \eqn{\lambda}, compute the unique value of \eqn{y}
minimizing
\deqn{\frac{1}{2} \Vert x - y \Vert_2^2 +
      \sum_{i=1}^n \lambda_i |y|_{(i)}.}
}
\details{
At present, two methods for computing the sorted L1 prox are
supported. By default, we use a fast custom C implementation. Since SLOPE
can be viewed as an isotonic regression problem, the prox can also be
computed using the \code{isotone} package. This option is provided
primarily for testing.
}
