% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABSLOPE.R
\name{rescale_all}
\alias{rescale_all}
\title{rescale_all}
\usage{
rescale_all(results, x_miss)
}
\arguments{
\item{results}{Design matrix}

\item{x_miss}{Response vector}
}
\description{
rescale_all
}
\keyword{internal}
