% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roughness_index.R
\name{roughness_index}
\alias{roughness_index}
\title{Roughness Index}
\usage{
roughness_index(x, sp_range)
}
\arguments{
\item{x}{Input DEM raster}

\item{w}{w Number of cells for the window size}
}
\description{
Calculates Roughness Index, based in Trevisani and Cavalli (2016).
}
\examples{
\dontrun{
library(terra)
library(slope)
dem <- terra::rast(system.file("ex/elev.tif", package = "terra"))
ri <- slope::roughness_index(dem, 5)
plot(ri)
}
}
