% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roughness_sediment_connectivity.R
\name{roughness_sediment_connectivity}
\alias{roughness_sediment_connectivity}
\title{Roughness Sediment Connectivity}
\usage{
roughness_sediment_connectivity(x, w, flow_acc, flow_path)
}
\arguments{
\item{x}{Input DEM raster}

\item{w}{Number of cells for the window size}

\item{flow_acc}{Input Flow Accumulation raster}

\item{flow_path}{Input Flow Path raster}
}
\description{
Calculates sediment connectivity, based in Cavalli et al. (2013).
}
\examples{
\dontrun{
library(terra)
library(slope)
dem <- terra:rast(system.file("ex/elev.tif", package = "terra"))
flow_acc <- terra:rast(system.file("ex/accum.tif", package = "terra"))
flow_path <- terra:rast(system.file("ex/flow_path.tif", package = "terra"))
rsc <- slope::roughness_sediment_connectivity(dem, 5, fa, fp)
plot(rsc)
}
}
