% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sediment_connectivity.R
\name{sediment_connectivity}
\alias{sediment_connectivity}
\title{Sediment Connectivity}
\usage{
sediment_connectivity(x, c, w, flow_acc, flow_path)
}
\arguments{
\item{x}{Input DEM raster}

\item{c}{C Factor reclassified based in Revised Equation Soil Loss Equation}

\item{w}{Number of cells for the window size}

\item{flow_acc}{Input Flow Accumulation raster}

\item{flow_path}{Input Flow Path raster}
}
\description{
Calculates sediment connectivity, based in Borelli et al. (2008).
}
\examples{
\dontrun{
library(terra)
library(slope)
dem <- terra:rast(system.file("ex/elev.tif", package = "terra"))
c <- terra::rast(system.file("ex/lulc.tif", package = "terra"))
flow_acc <- terra:rast(system.file("ex/accum.tif", package = "terra"))
flow_path <- terra:rast(system.file("ex/flow_path.tif", package = "terra"))
sc <- slope::sediment_connectivity(dem, 5, fa, fp)
plot(sc)
}
}
