% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sediment_transport_index.R
\name{sediment_transport_index}
\alias{sediment_transport_index}
\title{Sediment Transport Index}
\usage{
sediment_transport_index(x, type, m, n)
}
\arguments{
\item{x}{Input DEM raster file.}

\item{type}{Float. 'cells', 'sca' and 'ca'.}

\item{m}{Float. Exponent of first equation.}

\item{n}{Float. Exponent of second equation.}
}
\description{
Calculates Sediment Transport Index, based in Moore and Burch (1986).
}
\examples{
library(terra)
library(slope)
dem <- terra:rast(system.file("ex/elev.tif", package = "terra"))
twi <- slope::topographic_wetness_index(dem, 'sca')
plot(twi)
}
