% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_diagnostics.R
\name{setup_diagnostics}
\alias{setup_diagnostics}
\title{Setup a data.frame of diagnostics}
\usage{
setup_diagnostics(res)
}
\arguments{
\item{res}{the result from calling the C++ routine used to fit a model
in SLOPE}
}
\value{
A data.frame
}
\description{
Setup a data.frame of diagnostics
}
\keyword{internal}
