% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupDiagnostics.R
\name{setupDiagnostics}
\alias{setupDiagnostics}
\title{Setup a data.frame of diagnostics}
\usage{
setupDiagnostics(res)
}
\arguments{
\item{res}{the result from calling the C++ routine used to fit a model
in SLOPE}
}
\value{
A data.frame
}
\description{
Setup a data.frame of diagnostics
}
\keyword{internal}
