% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLOBE.R
\name{SLOBEC}
\alias{SLOBEC}
\title{SLOBEC}
\usage{
SLOBEC(
  start,
  Xmis,
  Xinit,
  Y,
  a_prior,
  b_prior,
  Covmat = diag(rep(1, length(start))),
  sigma = 1,
  FDR = 0.05,
  tol = 1e-04,
  known_sigma = FALSE,
  max_iter = 100L,
  verbose = FALSE,
  BH = TRUE,
  known_cov = FALSE
)
}
\arguments{
\item{start}{words, words}

\item{Xmis}{A number}
}
\description{
SLOBEC
}
