% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLOPE-package.R
\docType{package}
\name{SLOPE-package}
\alias{SLOPE-package}
\title{SLOPE: Sorted L1 Penalized Estimation}
\description{
Efficient implementations for Sorted L-One Penalized Estimation (SLOPE): generalized linear models regularized with the sorted L1-norm (Bogdan et al. 2015). Supported models include ordinary least-squares regression, binomial regression, multinomial regression, and Poisson regression. Both dense and sparse predictor matrices are supported. In addition, the package features predictor screening rules that enable fast and efficient solutions to high-dimensional problems.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jolars.github.io/SLOPE/}
  \item \url{https://github.com/jolars/SLOPE}
  \item Report bugs at \url{https://github.com/jolars/SLOPE/issues}
}

}
\keyword{internal}
