% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopeforms.R
\name{slopeforms}
\alias{slopeforms}
\title{Slope Forms}
\usage{
slopeforms(x, sp_range, crs)
}
\arguments{
\item{x}{Input DEM raster file.}

\item{sp_range}{Numeric. Number of neighbor cells for multiscalar analysis.}

\item{crs}{Input geodesic reference systems.}
}
\description{
Detects the nine landform elements based on Dikau (1989).
}
\details{
Each code refers to a specific pattern of slope forms:
\itemize{
  \item \code{11}: Concave-Concave slope (hollows).
  \item \code{12}: Concave-Concave slope.
  \item \code{13}: Concave-Convex slope.
  \item \code{21}: Concave-Straight slope.
  \item \code{22}: Straight-Straight slope.
  \item \code{23}: Convex-Straight slope.
  \item \code{31}: Concave-Convex slope.
  \item \code{32}: Straight-Convex slope.
  \item \code{33}: Convex-Convex slope.
}
}
\examples{
library(terra)
library(slope)
dem <- terra::rast(system.file("ex/elev.tif", package = "terra"))
sf <- slope::slopeforms(dem, 7, 'EPSG:5880')

}
