% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams.R
\name{streams}
\alias{streams}
\title{\strong Streams}
\usage{
streams(x, threshold)
}
\arguments{
\item{x}{Input Flow Accumulation raster file.}

\item{threshold}{Numeric. For extraction of streams.}
}
\description{
Performs identification of streams in DEMs based on the wbt_extract_streams() function (Lindsay, 2016) for calculates sediment connectivity (Cavalli et al. 2013).
}
\examples{
library(terra)
library(slope)
fa <- terra::rast(system.file("ex/accum.tif", package = "terra"))
s <- slope::streams(fa, 500)
plot(s)
}
