% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topographic_wetness_index.R
\name{topographic_wetness_index}
\alias{topographic_wetness_index}
\title{Topographic Wetness Index}
\usage{
topographic_wetness_index(x)
}
\arguments{
\item{x}{Input DEM raster file.}

\item{type}{Float. 'cells', 'sca' and 'ca'.}
}
\description{
Calculates Topographic Wetness Index, based in Beven and Kirbky (1979).
}
\examples{
library(terra)
library(slope)
dem <- terra:rast(system.file("ex/elev.tif", package = "terra"))
twi <- slope::topographic_wetness_index(dem, 'sca')
plot(twi)
}
