% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformed_slope.R
\name{transformed_slope}
\alias{transformed_slope}
\title{Transformation (Normalization) Slope}
\usage{
transformed_slope(x, w, lambda, crs, type)
}
\arguments{
\item{x}{Input DEM raster}

\item{w}{Number of cells for the window size}

\item{lambda}{Only type [1] and different from zero}

\item{crs}{Coordinate reference system (e.g., 'EPSG:5880')}

\item{type}{Normalization methods based on Evans (1977, 2015) [0] or Csillik et al. (2015) [1].}
}
\description{
Slope normalization based in Csillik et al. (2015)
}
\examples{
\dontrun{
library(terra)
library(slope)
dem <- terra:rast(system.file("ex/elev.tif", package = "terra"))
tc <- slope::transformed_slope(dem, 7, 0.1, 'EPSG:5880', 1)
plot(tc)
}
}
