% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wine}
\alias{wine}
\title{Wine cultivars}
\format{
178 observations from 13 variables represented as a list consisting
of a categorical response vector \code{y}
with three levels: \emph{A}, \emph{B}, and \emph{C} representing different
cultivars of wine as well as \code{x}: a sparse feature matrix of class
'dgCMatrix' with the following variables:
\describe{
\item{alcohol}{alcoholic content}
\item{malic}{malic acid}
\item{ash}{ash}
\item{alcalinity}{alcalinity of ash}
\item{magnesium}{magnemium}
\item{phenols}{total phenols}
\item{flavanoids}{flavanoids}
\item{nonflavanoids}{nonflavanoid phenols}
\item{proanthocyanins}{proanthocyanins}
\item{color}{color intensity}
\item{hue}{hue}
\item{dilution}{OD280/OD315 of diluted wines}
\item{proline}{proline}
}
}
\source{
Dua, D. and Karra Taniskidou, E. (2017). UCI Machine Learning Repository
\url{http://archive.ics.uci.edu/ml}. Irvine, CA: University of California,
School of Information and Computer Science.

\url{https://raw.githubusercontent.com/hadley/rminds/master/1-data/wine.csv}

\url{https://www.csie.ntu.edu.tw/~cjlin/libsvmtools/datasets/multiclass.html#wine}
}
\usage{
wine
}
\description{
A data set of results from chemical analysis of wines grown in Italy
from three different cultivars.
}
\keyword{datasets}
