% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance.R
\name{deviance.SLOPE}
\alias{deviance.SLOPE}
\title{Model Deviance}
\usage{
\method{deviance}{SLOPE}(object, ...)
}
\arguments{
\item{object}{an object of class \code{'SLOPE'}.}

\item{...}{ignored}
}
\value{
For Gaussian models this is twice the residual sums of squares. For
all other models, two times the negative loglikelihood is returned.
}
\description{
Model Deviance
}
\examples{
fit <- SLOPE(heart$x, heart$y, family = "binomial")
deviance(fit)
}
\seealso{
\code{\link[=SLOPE]{SLOPE()}}

Other SLOPE-methods: 
\code{\link{coef.SLOPE}()},
\code{\link{plot.SLOPE}()},
\code{\link{predict.SLOPE}()},
\code{\link{print.SLOPE}()},
\code{\link{score}()},
\code{\link{summary.SLOPE}()}
}
\concept{SLOPE-methods}
