% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissection_index.R
\name{dissection_index}
\alias{dissection_index}
\title{Dissection Index}
\usage{
dissection_index(x, w)
}
\arguments{
\item{x}{Input DEM raster}

\item{w}{Number of cells for the window size}
}
\description{
Calculates the Dissection Index (DI) to detect spatial patterns of terrain dissection, based on Evans (1972).
}
\examples{
\dontrun{
library(terra)
library(slope)
dem <- terra:rast(system.file("ex/elev.tif", package = "terra"))
di <- slope::dissection_index(dem, 7)
plot(di)
}
}
