% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissection_index.R
\name{dissection_index}
\alias{dissection_index}
\title{\strong Dissection Index}
\usage{
dissection_index(x, sp_range)
}
\arguments{
\item{x}{Input DEM raster file.}

\item{sp_range}{Numeric. Number of neighbor cells for multiscalar analysis.}
}
\description{
Detects spatial patterns of dissection, based in Evans (1972).
}
\examples{
library(terra)
library(slope)
dem <- terra:rast(system.file("ex/elev.shp", package = "terra"))
di <- slope::dissection_index(dem, 7)
plot(di)
}
