% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevr.R
\name{elevr}
\alias{elevr}
\title{Elevr}
\usage{
elevr(dem, aoi, api_key, output_file)
}
\arguments{
\item{dem}{Currently supports: "SRTMGL3", "SRTMGL1", "SRTMGL1_E", "AW3D30", "AW3D30_E",
"SRTM15Plus", "NASADEM", "COP30", "COP90", "EU_DTM", "GEDI_L3", "GEBCOIceTopo", "GEBCOSubIceTopo",
"CA_MRDEM_DTM", "CA_MRDEM_DSM", "ANADEM", "GEDTM30" from the OpenTopography API global datasets.}

\item{aoi}{Defines the area of interest to crop/bound the elevation data, in *sf* format}

\item{api_key}{OpenTopography API key}

\item{output_file}{File path to save the elevation data in raster format}
}
\description{
This function provides access to global raster elevation data from the OpenTopography API.
}
\examples{
\dontrun{
library(sf)
library(slope)
aoi <- sf::read_sf('ex/aoi.shp')
dem <- slope::elevr('GEDTM30', aoi, api_key, 'ex/dem.tif')
plot(dem)
}
}
