% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.R
\name{fill}
\alias{fill}
\title{\strong Spurious Depressions Filling}
\usage{
fill(x, y)
}
\arguments{
\item{x}{Input DEM raster file.}

\item{y}{Output filled raster file.}
}
\description{
Performs the filling of spurious depressions in DEMs,
\code based on the wbt_fill_depressions() function (Lindsay, 2016)
}
\examples{
library(terra)
dem <- terra::rast(system.file("ex/elev.tif", package="terra"))
dem_fill <- slope::fill(dem, 'fill.tif')
plot(dem_fill)
}
