% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.R
\name{fill}
\alias{fill}
\title{Fill}
\usage{
fill(x, y)
}
\arguments{
\item{x}{Input DEM raster}

\item{y}{Output filled raster file}
}
\description{
Performs the filling of spurious depressions in DEMs based on the wbt_fill_depressions() function (Lindsay, 2016).
}
