% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_accumulation.R
\name{flow_accumulation}
\alias{flow_accumulation}
\title{Flow Accumulation}
\usage{
flow_accumulation(x, type)
}
\arguments{
\item{x}{Input DEM raster}

\item{type}{Flow accumulation calculation type: 'cells', 'sca', or 'ca'}
}
\description{
Performs the flow accumulation in DEMs based on the wbt_flow_accumulation_full_workflow() function (Lindsay, 2016) for calculates sediment connectivity (Cavalli et al. 2013).
}
\examples{
\dontrun{
library(terra)
library(slope)
dem <- terra::rast(system.file("ex/elev.tif", package = "terra"))
fa <- slope::flow_accumulation(dem, 'sca')
plot(fa)
}
}
