% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcd.R
\name{geomorphic_change_detection}
\alias{geomorphic_change_detection}
\title{Geomorphic Change Detection}
\usage{
geomorphic_change_detection(z_actual, z_dem, crs, type)
}
\arguments{
\item{z_actual}{Input DEM raster in second moment in time, using the same Earth Gravitational Model (EGM) for reduction in vertical error.}

\item{z_dem}{Input DEM raster in first moment in time, using the same Earth Gravitational Model (EGM) for reduction in vertical error.}

\item{crs}{Coordinate reference system (e.g., 'EPSG:5880')}

\item{type}{Choice between deposition [0] or erosion [1]}
}
\description{
Detects geomorphic change detection (DEMs of Difference) based in Wheathon et al. (2010).
}
\examples{
\dontrun{
library(terra)
z_actual <- terra::rast(system.file("ex/elev.tif", package = "terra"))
z_dem <- terra::rast(system.file("ex/elev.tif", package = "terra"))
gcd <- slope::geomorphic_change_detection(z_actual, z_dem, 'EPSG:5880', 1)
plot(gcd)
}
}
