% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillshade.R
\name{hillshade}
\alias{hillshade}
\title{\strong Hillshade for slope patterns visualization}
\usage{
hillshade(x, w, angle, direction)
}
\arguments{
\item{x}{Input slopeforms raster result.}

\item{w}{Numeric. Moving window size for the median filter.}

\item{angle}{Numeric. Illumination angle (altitude) in degrees.}

\item{direction}{Numeric. Illumination direction (azimuth) in degrees.}
}
\description{
Calculates the area and percentages of slope forms for tabular visualization.
}
\examples{
library(terra)
dem <- terra::rast(system.file("ex/elev.tif", package="terra"))
shade <- slope::hillshade(dem, 3, 45, 315)
plot(shade)
}
