% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mininum_mappable_area.R
\name{mininum_mappable_area}
\alias{mininum_mappable_area}
\title{\strong Minimum Mappable Area}
\usage{
mininum_mappable_area(scale, res)
}
\arguments{
\item{scale}{Numeric. Cartographic scale}

\item{res}{Numeric. Resolution of DEM.}
}
\description{
Calculates minimum mappable area.
}
\examples{
library(terra)
library(slope)
dem <- terra::rast(system.file("ex/elev.tif", package="terra"))
print(dem)
mma <- slope::minimum_mappable_area(100000, 30)
}
