% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_occurrence_samples.R
\name{non_occurence_samples}
\alias{non_occurence_samples}
\title{Non Occurrence Samples}
\usage{
non_occurence_samples(x, y, distance, n_samples, drop_cols_non, drop_cols, col)
}
\arguments{
\item{x}{Input stack raster}

\item{y}{Input samples (vector files)}

\item{distance}{Area or zone of influence of the event occurrence samples.}

\item{n_samples}{Regarding the number of samples representing the non-occurrence of events, it is suggested, in order to balance the sample set, to use a sample size equal to the number of event occurrences.}

\item{drop_cols_non}{This allows you to remove information that will not be used; it is suggested to leave only the column referring to the non-occurrence of events, e.g [0].}

\item{drop_cols}{It allows for the removal of unused information; it is suggested to retain only the column regarding event occurrences, e.g [1].}

\item{col}{Target field of occurrence.}
}
\description{
Determines non-occurrence samples of susceptibility based on the Buffer Controlling Samples (BCS) method described by Gu et al. (2024).
}
\examples{
\dontrun{
library(pacman)
p_load(terra, sf, slope)
rasters <- terra::rast(system.file("ex/stack.tif", package = "terra"))
samples <- terra::vect(system.file("ex/samples.shp", package = "terra"))
non_occurrence <- slope::non_occurrence_samples(rasters, samples, 500, 1218, -c(1:14), -c(1), 'classes')
plot(non_occurrence)
}
}
