% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalization}
\alias{normalization}
\title{\strong Normalization}
\usage{
normalization(x, sp_range)
}
\arguments{
\item{x}{Input DEM raster file.}

\item{sp_range}{Numeric. Number of neighbor cells for multiscalar analysis.}
}
\description{
Performs normalization of Land Surface Parameters (LSPs), based Huang et al. (2021)
}
\examples{
library(terra)
library(slope)
dem <- terra:rast(system.file("ex/elev.tif", package = "terra"))
ni <- slope::normalization(dem, 7)
plot(ni)
}
