% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physical_geomorphometry.R
\name{physical_geomorphometry}
\alias{physical_geomorphometry}
\title{Physical Geomorphometry}
\usage{
physical_geomorphometry(x)
}
\arguments{
\item{x}{Input slopeforms raster result.}
}
\description{
\strong{Physical Geomorphometry}
}
\details{
Calculates the area and percentages of slope forms for tabular visualization.
}
\examples{
\dontrun{
library(terra)
library(slope)
sf <- terra::rast('ex/slopeforms.tif', package = "terra")
pg <- slope::physical_geomorphometry(sf)
plot(pg)
}
}
