% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physical_geomorphometry_plot.R
\name{physical_geomorphometry_plot}
\alias{physical_geomorphometry_plot}
\title{\strong Physical Geomorphometry in Graphic}
\usage{
physical_geomorphometry_plot(x)
}
\arguments{
\item{x}{Input slopeforms raster result.}
}
\description{
Calculates the percentages of slope forms for graphical visualization.
}
\examples{
library(terra)
library(slope)
sf <- terra::rast('slopeforms.tif')
pgv <- slope::physical_geomorphometry_plot(sf)
plot(pgv)
}
