% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rci.R
\name{rci}
\alias{rci}
\title{Roughness Concentration Index}
\usage{
rci(x, sp_range, crs, aoi, unit)
}
\arguments{
\item{x}{Input DEM raster file.}

\item{sp_range}{Numeric. Number of neighbor cells for multiscalar analysis.}

\item{crs}{Input geodesic reference systems.}

\item{aoi}{Area of interest in spatvect.}

\item{unit}{Metrics units in square kilometers.}
}
\description{
Calculates the Roughness Concentration Index (RCI) based on the methodology of Sampaio and Augustin (2014).
}
\examples{
library(terra)
library(slope)
aoi <- terra:vect(system.file("ex/aoi.shp", package = "terra"))
dem <- terra:rast(system.file("ex/elev.tif", package = "terra"))
r <- slope::rci(dem, 300, 'EPSG:5880', aoi, 1000000)
plot(r)
}
