% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rci.R
\name{rci}
\alias{rci}
\title{Roughness Concentration Index}
\usage{
rci(x, w, crs, aoi, unit)
}
\arguments{
\item{x}{Input DEM raster file.}

\item{w}{Number of cells for the window size}

\item{crs}{Coordinate reference system (e.g., 'EPSG:5880')}

\item{aoi}{Area of interest}

\item{unit}{Metrics units in square kilometers.}
}
\description{
Calculates the Roughness Concentration Index (RCI) based on the methodology of Sampaio and Augustin (2014).
}
\examples{
\dontrun{
library(terra)
library(slope)
aoi <- terra:vect(system.file("ex/aoi.shp", package = "terra"))
dem <- terra:rast(system.file("ex/elev.tif", package = "terra"))
rci <- slope::rci(dem, 300, 'EPSG:5880', aoi, 1000000)
plot(rci)
}
}
