% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridges_tops.R
\name{ridges_tops}
\alias{ridges_tops}
\title{Ridges and Tops}
\usage{
ridges_tops(x, w, type)
}
\arguments{
\item{x}{Input DEM raster}

\item{w}{Number of cells for the window size}

\item{type}{Each number represents a specific landform, for example: (1) Convex hilltops and Interfluves, (2) Sharp crests, and (3) Ridges}
}
\description{
Detects different types of ridges and tops, based on Silveira and Silveira (2020).
The primary difference is the use of a rectangular local neighborhood shape instead
of a circular one, in the calculation of the White Top Hat (WTH) (Rodriguez et al., 2002).
For more satisfactory results, it is recommended to define the
moving window based on the minimum mappable area.
}
\examples{
\dontrun{
library(terra)
library(slope)
dem <- terra:rast(system.file("ex/elev.tif", package = "terra"))
r <- slope::ridges_tops(dem, 7, 1)
plot(r)
}
}
