% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridges_tops.R
\name{ridges_tops}
\alias{ridges_tops}
\title{Ridges and Tops}
\usage{
ridges_tops(x, sp_range, type)
}
\arguments{
\item{x}{Input DEM raster file.}

\item{sp_range}{Numeric. Number of neighbor cells for multiscalar analysis.}

\item{type}{Numeric. Each number represents a specific landform, for example: (1) Convex Hilltops and Interfluves, (2) Sharp Crests, and (3) Ridges}
}
\description{
\strong{Ridges and Tops}
}
\details{
Detects different types of Ridges and Tops, based on Silveira and Silveira (2020).
\code The primary difference is the use of a rectangular local neighborhood shape instead
\code of a circular one, in calculus of White Top Hat (WTH) (Rodriguez et al. 2002).
\code For more satisfactory results, it is recommended to define the
\code moving window based on the minimum mappable area.
}
\examples{
\dontrun{
library(terra)
library(slope)
dem <- terra:rast(system.file("ex/elev.tif", package = "terra"))
r <- slope::ridges_tops(dem, 7, 1)
plot(r)
}
}
