% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roughness_sediment_connectivity.R
\name{roughness_sediment_connectivity}
\alias{roughness_sediment_connectivity}
\title{Roughness Sediment Connectivity}
\usage{
roughness_sediment_connectivity(x, sp_range, fa, fp)
}
\arguments{
\item{x}{Input DEM raster file.}

\item{sp_range}{Numeric. Number of neighbor cells for multiscalar analysis.}

\item{fa}{Input Flow Accumulation raster file.}

\item{fp}{Input Flow Path raster file.}
}
\description{
Calculates sediment connectivity where w is roughness, based in Cavalli et al. (2013).
}
\examples{
library(terra)
library(slope)
dem <- terra:rast(system.file("ex/elev.tif", package = "terra"))
fa <- terra:rast(system.file("ex/accum.tif", package = "terra"))
fp <- terra:rast(system.file("ex/flow_path.tif", package = "terra"))
ic <- slope::roughness_sediment_connectivity(dem, 5, fa, fp)
plot(ic)
}
