% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sediment_connectivity.R
\name{sediment_connectivity}
\alias{sediment_connectivity}
\title{Sediment Connectivity}
\usage{
sediment_connectivity(x, sp_range, fa, fp)
}
\arguments{
\item{x}{Input DEM raster file.}

\item{sp_range}{Numeric. Number of neighbor cells for multiscalar analysis.}

\item{fa}{Input Flow Accumulation raster file.}

\item{fp}{Input Flow Path raster file.}
}
\description{
\strong{Sediment Connectivity}
}
\details{
Calculates sediment connectivity, based in Cavalli et al. (2013)
}
\examples{
\dontrun{
library(terra)
library(slope)
dem <- terra:rast(system.file("ex/elev.tif", package = "terra"))
fa <- terra:rast(system.file("ex/accum.tif", package = "terra"))
fp <- terra:rast(system.file("ex/flow_path.tif", package = "terra"))
ic <- slope::sediment_connectivity(dem, 7, fa, fp)
plot(ic)
}
}
