% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sediment_transport_index.R
\name{sediment_transport_index}
\alias{sediment_transport_index}
\title{Sediment Transport Index}
\usage{
sediment_transport_index(x, type, crs, m, n, res)
}
\arguments{
\item{x}{Input DEM raster}

\item{type}{Flow accumulation calculation type: 'cells', 'sca', or 'ca'}

\item{crs}{Coordinate reference system (e.g., 'EPSG:5880')}

\item{m}{Exponent of first equation}

\item{n}{Exponent of second equation}

\item{res}{Resolution of DEM}
}
\description{
Calculates Sediment Transport Index, based in Moore and Burch (1986).
}
\examples{
\dontrun{
library(terra)
library(slope)
dem <- terra:rast(system.file("ex/elev.tif", package = "terra"))
sti <- slope::topographic_wetness_index(dem, 'sca')
plot(sti)
}
}
