% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slider.R
\name{slider}
\alias{slider}
\title{\strong Debris Flow Prediction}
\usage{
slider(x, sp_range, crs, less_int_slope, more_int_slope, c)
}
\arguments{
\item{x}{Input DEM raster file.}

\item{sp_range}{Numeric. Number of neighbor cells for multiscalar analysis.}

\item{crs}{Input geodesic reference systems.}

\item{less_int_slope}{Numeric. Minimum slope angle.}

\item{more_int_slope}{Numeric. Maximum slope angle.}

\item{c}{Concave-concave slope (hollows), referring to code 11, see details in slopeforms.}
}
\description{
Detects debris flow based on the identification of hollows and critical slope thresholds.
}
\examples{
library(terra)
dem <- terra::rast(system.file("ex/elev.tif", package="terra"))
sd <- slope::slider(dem, 7, 'EPSG:5880', 21, 25, 11)
plot(sd)
}
