% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slider.R
\name{slider}
\alias{slider}
\title{Slider}
\usage{
slider(x, sp_range, crs, less_int_slope, more_int_slope, c)
}
\arguments{
\item{x}{Input DEM raster}

\item{crs}{Coordinate reference system (e.g., 'EPSG:5880')}

\item{w}{Number of cells for the window size}

\item{min_slope}{Minimum slope angle}

\item{max_slope}{Maximum slope angle}

\item{type}{Concave-concave slope (hollows), referring to code 11, see details in slopeforms}
}
\description{
Detects debris flow based on the identification of hollows and critical slope thresholds.
}
\examples{
\dontrun{
library(terra)
library(slope)
dem <- terra::rast(system.file("ex/elev.tif", package = "terra"))
slider <- slope::slider(dem, 7, 'EPSG:5880', 21, 25, 11)
plot(slider)
}
}
