% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLOPE-package.R
\docType{package}
\name{SLOPE-package}
\alias{SLOPE-package}
\alias{_PACKAGE}
\title{SLOPE: Sorted L1 Penalized Estimation}
\description{
Efficient implementations for Sorted L-One Penalized Estimation (SLOPE): generalized linear models regularized with the sorted L1-norm (Bogdan et al. (2015) <doi:10/gfgwzt>). Supported models include ordinary least-squares regression, binomial regression, multinomial regression, and Poisson regression. Both dense and sparse predictor matrices are supported. In addition, the package features predictor screening rules that enable fast and efficient solutions to high-dimensional problems.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jolars.github.io/SLOPE/}
  \item \url{https://github.com/jolars/SLOPE}
  \item Report bugs at \url{https://github.com/jolars/SLOPE/issues}
}

}
\author{
\strong{Maintainer}: Johan Larsson \email{johan.larsson@stat.lu.se} (\href{https://orcid.org/0000-0002-4029-5945}{ORCID})

Authors:
\itemize{
  \item Jonas Wallin \email{jonas.wallin@stat.lu.se} (\href{https://orcid.org/0000-0003-0381-6593}{ORCID})
  \item Malgorzata Bogdan
  \item Ewout van den Berg
  \item Chiara Sabatti
  \item Emmanuel Candes
  \item Evan Patterson
  \item Weijie Su
}

Other contributors:
\itemize{
  \item Jerome Friedman (code adapted from 'glmnet') [contributor]
  \item Trevor Hastie (code adapted from 'glmnet') [contributor]
  \item Rob Tibshirani (code adapted from 'glmnet') [contributor]
  \item Balasubramanian Narasimhan (code adapted from 'glmnet') [contributor]
  \item Noah Simon (code adapted from 'glmnet') [contributor]
  \item Junyang Qian (code adapted from 'glmnet') [contributor]
  \item Akarsh Goyal [contributor]
  \item Jakub Kała [contributor]
  \item Krystyna Grzesiak [contributor]
}

}
\keyword{internal}
