% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopeforms.R
\name{slopeforms}
\alias{slopeforms}
\title{Slope Forms}
\usage{
slopeforms(x, w, crs)
}
\arguments{
\item{x}{Input DEM raster file.}

\item{w}{Number of cells for the window size}

\item{crs}{Coordinate reference system (e.g., 'EPSG:5880')}
}
\description{
Detects the nine landform elements based on Dikau (1989).
}
\details{
Each code refers to a specific pattern of slope forms:
\itemize{
  \item \code{11}: Concave-Concave slope (hollows).
  \item \code{12}: Concave-Concave slope.
  \item \code{13}: Concave-Convex slope.
  \item \code{21}: Concave-Straight slope.
  \item \code{22}: Straight-Straight slope.
  \item \code{23}: Convex-Straight slope.
  \item \code{31}: Concave-Convex slope.
  \item \code{32}: Straight-Convex slope.
  \item \code{33}: Convex-Convex slope.
}
}
\examples{
\dontrun{
library(terra)
library(slope)
dem <- terra::rast(system.file("ex/elev.tif", package = "terra"))
slope_forms <- slope::slopeforms(dem, 7, 'EPSG:5880')
plot(slope_forms)
}
}
