% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streams.R
\name{streams}
\alias{streams}
\title{Streams}
\usage{
streams(x, threshold)
}
\arguments{
\item{x}{Input Flow Accumulation raster}

\item{threshold}{Threshold value for stream extraction}
}
\description{
Identification of streams in DEMs based on the wbt_extract_streams() function (Lindsay, 2016) to calculate sediment connectivity (Cavalli et al., 2013).
}
\examples{
\dontrun{
library(terra)
library(slope)
flow_acc <- terra::rast(system.file("ex/accum.tif", package = "terra"))
streams <- slope::streams(flow_acc, 500)
plot(streams)
}
}
