% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_relief_ratio.R
\name{surface_relief_ratio}
\alias{surface_relief_ratio}
\title{Surface Relief Ratio}
\usage{
surface_relief_ratio(x, w)
}
\arguments{
\item{x}{Input DEM raster}

\item{w}{Number of cells for the window size}
}
\description{
Calculates Surface Relief Ratio, based on Berry (2002).
}
\examples{
\dontrun{
library(terra)
library(slope)
dem <- terra:rast(system.file("ex/elev.tif", package = "terra"))
srr <- slope::surface_relief_ratio(dem, 7)
plot(srr)
}
}
