% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_relief_ratio.R
\name{surface_relief_ratio}
\alias{surface_relief_ratio}
\title{\strong Surface Relief Ratio}
\usage{
surface_relief_ratio(x, sp_range)
}
\arguments{
\item{x}{Input DEM raster file.}

\item{sp_range}{Numeric. Number of neighbor cells for multiscalar analysis.}
}
\description{
Calculates Surface Relief Ratio, based Berry (2002).
}
\examples{
library(terra)
library(slope)
dem <- terra:rast(system.file("ex/elev.tif", package = "terra"))
srr <- slope::surface_relief_ratio(dem, 7)
plot(srr)
}
