% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformed_curvature.R
\name{transformed_curvature}
\alias{transformed_curvature}
\title{Transformation (Normalization) Curvature}
\usage{
transformed_curvature(x, w, k, crs, type)
}
\arguments{
\item{x}{Input DEM raster}

\item{w}{Number of cells for the window size}

\item{k}{Kurtosis close to 0}

\item{crs}{Coordinate reference system (e.g., 'EPSG:5880')}

\item{type}{Profile curvature [0] or Plan curvature [1]}
}
\description{
Curvature normalization based in Evans (1972) and Csillik et al. (2015)
}
\examples{
\dontrun{
library(terra)
library(slope)
dem <- terra:rast(system.file("ex/elev.tif", package = "terra"))
tc <- slope::transformed_curvature(dem, 7, 0.1, 'EPSG:5880', 1)
plot(tc)
}
}
