% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformed_curvature.R
\name{transformed_curvature}
\alias{transformed_curvature}
\title{Transformation (Normalization) Curvature}
\usage{
transformed_curvature(x, sp_range, k, crs, type)
}
\arguments{
\item{x}{Input DEM raster file.}

\item{sp_range}{Numeric. Number of neighbor cells for multiscalar analysis.}

\item{k}{Numeric. Kurtosis close to 0.}

\item{crs}{Numeric. Input geodesic reference systems.}

\item{type}{Numeric. Choice between Profile Curvature [0] or Plan Curvature [1].}
}
\description{
\strong{Transformation (Normalization) Curvature}
}
\details{
Curvature normalization, based in Evans (1972) and Csillik et al. (2015)
}
\examples{
\dontrun{
library(terra)
library(slope)
dem <- terra:rast(system.file("ex/elev.tif", package = "terra"))
tc <- slope::transformed_curvature(dem, 7, 0.1, 'EPSG:5880', 1)
plot(tc)
}
}
