% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformed_slope.R
\name{transformed_slope}
\alias{transformed_slope}
\title{Transformation (Normalization) Slope}
\usage{
transformed_slope(x, sp_range, lambda, crs, type)
}
\arguments{
\item{x}{Input DEM raster file.}

\item{sp_range}{Numeric. Number of neighbor cells for multiscalar analysis.}

\item{lambda}{Numeric. Only type [1]}

\item{crs}{Input geodesic reference systems.}

\item{type}{Numeric. Choice between normalization based in Evans (1977) (2015) [0] or Csillik et al. (2015) [1].}
}
\description{
\strong{Transformation (Normalization) Slope}
}
\details{
Slope normalization, based in Csillik et al. (2015)
}
\examples{
\dontrun{
library(terra)
library(slope)
dem <- terra:rast(system.file("ex/elev.tif", package = "terra"))
tc <- slope::transformed_slope(dem, 7, 0.1, 'EPSG:5880', 1)
plot(tc)
}
}
