% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valleys.R
\name{valleys}
\alias{valleys}
\title{Valleys}
\usage{
valleys(x, w, type)
}
\arguments{
\item{x}{Input DEM raster}

\item{w}{Number of cells for the window size}

\item{type}{Each number represents a specific landform, for example: (1) Flat-bottomed valleys, (2) Open valleys, and (3) Incised valleys.}
}
\description{
Detects different types of valleys, based on Silveira and Silveira (2020).
The primary difference is the use of a rectangular local neighborhood shape instead
of a circular one, in the calculation of Black Top Hat (BTH) (Rodriguez et al., 2002).
For more satisfactory results, it is recommended to define the
moving window based on the minimum mappable area and the fill() function.
}
