% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valleys.R
\name{valleys}
\alias{valleys}
\title{Valleys}
\usage{
valleys(x, sp_range, c)
}
\arguments{
\item{x}{Input DEM raster file.}

\item{sp_range}{Numeric. Number of neighbor cells for multiscalar analysis.}

\item{type}{Numeric. Each number represents a specific landform, for example: (1) Flat-bottomed Valleys, (2) Open Valleys, and (3) Incised Valleys}
}
\description{
Detects different types of valleys, based on Silveira and Silveira (2020).
The primary difference is the use of a rectangular local neighborhood shape instead
of a circular one, in calculus of Black Top Hat (BTW) (Rodriguez et al. 2002).
For more satisfactory results, it is recommended to define the
moving window based on the minimum mappable area and function fill()
}
